% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{addMissingSpecColumns}
\alias{addMissingSpecColumns}
\title{Add missing columns for specification}
\usage{
addMissingSpecColumns(df, names)
}
\arguments{
\item{df}{data.frame data frame with data for the specification}

\item{names}{names of columns to include in new data.frame}
}
\description{
Add missing columns for specification
}
\details{
Takes in a data.frame and a set of column names,
returns a data.frame with the specified columns, assigning 0 to the
value of missing columns for all rows.
}
