% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{admm}
\alias{admm}
\title{Calculate quantile regression via admm estimation}
\usage{
admm(Y, X, est, vec, rho, tau, iterfun, k, err_u_thresh, verbose, ...)
}
\arguments{
\item{Y}{outcome variable}

\item{X}{variables to estimate}

\item{vec}{vector of parameters to update}

\item{rho}{stochastic relaxation parameter}

\item{tau}{quantile to estimate}

\item{iterfun}{function for updating parameters given new errors}

\item{k}{penalization parameter}

\item{err_u_thresh}{error threshold for convergence}

\item{verbose}{whether to be chatty}

\item{...}{other parameters to be passed to f}
}
\description{
Calculate quantile regression via admm estimation
}
