% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{all_match}
\alias{all_match}
\title{check if all values of two vectors match}
\usage{
all_match(x, y)

all_match(x, y)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\description{
check if all values of two vectors match

check if all values of two vectors match
}
