% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_effects.R
\name{bin_along_range}
\alias{bin_along_range}
\title{Bin a variable along a range}
\usage{
bin_along_range(x, size = NA, trim = 0.01)
}
\arguments{
\item{x}{variable to create binned sequence along}

\item{size}{size of bin}

\item{trim}{percent of obs to trim from both ends of variable}
}
\description{
Bin a variable along a range
}
