% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{bootstrapRows}
\alias{bootstrapRows}
\title{Function that fully bootstraps rows}
\usage{
bootstrapRows(rows, ...)

bootstrapRows(rows, ...)
}
\arguments{
\item{rows}{rows to potentially resample}

\item{...}{other arguments, ignored}
}
\description{
Function that fully bootstraps rows

Function that fully bootstraps rows
}
