% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{calc_gamma}
\alias{calc_gamma}
\title{Calculate Gamma}
\usage{
calc_gamma(
  Y,
  T_vec,
  J_vec,
  N,
  J,
  Time,
  Gamma = NULL,
  tau = 0.5,
  rho = 0.1,
  k = 1,
  err_u_thresh = 0.01,
  err_g_thresh = 0.01,
  verbose = F
)
}
\arguments{
\item{N}{number of assets}

\item{J}{integer representing number of groups}

\item{Time}{integer representing number of periods}

\item{tau}{quantile to calculate}

\item{rho}{stochastic relaxation parameter}

\item{k}{penalization parameter used in iteration, defaults to no penalization}

\item{err_u_thresh}{error threshold for u}

\item{err_g_thresh}{error threshold for g}

\item{DT}{data.table of values to be used in optimization}
}
\description{
Calculate Gamma
}
