% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{calc_r}
\alias{calc_r}
\title{Calculate r^{k+1} given parameters
Follows Yu and Lin (2017)}
\usage{
calc_r(rho, u, Y, Gamma, tau, One)
}
\arguments{
\item{rho}{relaxation parameter}

\item{u}{Lagrange multiplier term}
}
\description{
Calculate r^{k+1} given parameters
Follows Yu and Lin (2017)
}
