% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{calc_reduce_rank_gamma}
\alias{calc_reduce_rank_gamma}
\title{Iterate over K until getting a low enough rank for Gamma}
\usage{
calc_reduce_rank_gamma(
  Y,
  T_vec,
  J_vec,
  N,
  J,
  Time,
  Gamma = NULL,
  tau = 0.5,
  rho = 0.1,
  init_k = 2,
  desired_dim = NULL,
  err_u_thresh = 0.01,
  err_g_thresh = 0.01,
  verbose = F,
  maxiter = 100
)
}
\arguments{
\item{Y}{vector of outcomes}

\item{T_vec}{vector of time assignment}

\item{J_vec}{vector of group assignments}

\item{N}{number of assets}

\item{J}{integer representing number of groups}

\item{Time}{integer representing number of periods}

\item{tau}{quantile to calculate}

\item{rho}{stochastic relaxation parameter}

\item{err_u_thresh}{error threshold for u}

\item{err_g_thresh}{error threshold for g}

\item{k_init}{initial penalization parameter used in iteration, defaults to 1}
}
\description{
Iterate over K until getting a low enough rank for Gamma
}
