% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{calcAvgSpacing}
\alias{calcAvgSpacing}
\title{Computes means for various slices of a regression_spec by betas product.}
\usage{
calcAvgSpacing(x, betas, alpha, jstar, trim = 0.01)
}
\arguments{
\item{x}{Regression specification.}

\item{betas}{Fitted quantile coefficients for specification.}

\item{alpha}{Quantiles targeted.}

\item{jstar}{First quantile to be estimated (usually the center one)}

\item{trim}{Fraction (0 to 0.5) of observations to be trimmed from each end of
a given column in the regression_spec by betas product before the mean is computed.}
}
\value{
Matrix with rows containing means for each quantile for various slices
of the specification.
}
\description{
Computes means for various slices of a regression_spec by betas product.
}
