% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{cast_gamma}
\alias{cast_gamma}
\title{Reformat Gamma vector of quantile estimates for fixed effects}
\usage{
cast_gamma(Gamma, J_vec, T_vec)
}
\arguments{
\item{Gamma}{vector of coefficients for fixed effects}

\item{J_levels}{levels of grouping factor}

\item{T_levels}{levels of time factor}
}
\description{
Reformat Gamma vector of quantile estimates for fixed effects
}
