% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_se.R
\name{clusterSample}
\alias{clusterSample}
\title{Draws a subsample of clusters and (optionally) assigns
random, exponential weights to each cluster}
\usage{
clusterSample(
  cluster_indices,
  stratum_indices = NULL,
  M = 1,
  draw_weights = FALSE
)
}
\arguments{
\item{cluster_indices}{N_cluster x 2 vector with starting
and ending positions of each cluster.
NOTE that this requires that data are
sorted in this dimension}

\item{stratum_indices}{N_cluster x 1 vector with stratum indices
(factor variables). This enables the
researcher to stratify (e.g., on cluster size).
Set to NULL to take a random subsample.}

\item{M}{percentage of clusters to sample.  Must be between 0 and 1 (inclusive).}

\item{draw_weights}{If true, draw a vector of exponential weights to use in subsample}
}
\value{
list of
subsample_rows: Boolean vector of selected rows of sampled clusters
subsample_weights: either NULL, or vector of cluster weights
}
\description{
Draws a subsample of clusters and (optionally) assigns
random, exponential weights to each cluster
}
