% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.R
\name{coef.qs}
\alias{coef.qs}
\title{Method for getting coefficients from fitted qs model}
\usage{
\method{coef}{qs}(x, ...)
}
\arguments{
\item{x}{fitted qs model}

\item{...}{currently ignored}
}
\description{
Method for getting coefficients from fitted qs model
}
