% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{collapse_correctly}
\alias{collapse_correctly}
\title{helper function, collapse using correct method}
\usage{
collapse_correctly(x, l)
}
\arguments{
\item{x}{value,}

\item{l}{list}
}
\description{
helper function, collapse using correct method
}
