% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{columns_match_vector}
\alias{columns_match_vector}
\title{Check which rows of a matrix match a whole vector}
\usage{
columns_match_vector(X, Y)

columns_match_vector(X, Y)
}
\arguments{
\item{X}{first matrix}

\item{Y}{vector to match}
}
\description{
Check which rows of a matrix match a whole vector

Check which rows of a matrix match a whole vector
}
