% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{csrToDgc}
\alias{csrToDgc}
\title{Convert SparseM row compressed matrix to Matrix dgC matrix}
\usage{
csrToDgc(X)
}
\arguments{
\item{X}{column compressed matrix from Matrix package}
}
\description{
Convert SparseM row compressed matrix to Matrix dgC matrix
}
