% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_helpers.R
\name{cub_root}
\alias{cub_root}
\title{Find cube root of the form ax^3 + bx^2 + cx + d=0}
\usage{
cub_root(a, b, c, d)
}
\arguments{
\item{a}{parameter 1 in cubic equation}

\item{b}{parameter 2 in cubic equation}

\item{c}{parameter 3 in cubic equation}

\item{d}{parameter 4 in cubic equation}
}
\description{
Find cube root of the form ax^3 + bx^2 + cx + d=0
}
