% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_helpers.R
\name{cub_root_deriv}
\alias{cub_root_deriv}
\title{Derivative of a cubic root}
\usage{
cub_root_deriv(a, b, c, d)
}
\arguments{
\item{a}{parameter 1 in cubic equation}

\item{b}{parameter 2 in cubic equation}

\item{c}{parameter 3 in cubic equation}

\item{d}{parameter 4 in cubic equation}
}
\description{
Derivative of a cubic root
}
