% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_helpers.R
\name{cub_root_select_rconics}
\alias{cub_root_select_rconics}
\title{Calculate cubic roots using the Rconics package}
\usage{
cub_root_select_rconics(a, b, c, d, q_1, q_2)
}
\arguments{
\item{a}{parameter 1 in cubic equation}

\item{b}{parameter 2 in cubic equation}

\item{c}{parameter 3 in cubic equation}

\item{d}{parameter 4 in cubic equation}

\item{q_1}{lower quantile to bound the root}

\item{q_2}{upper quantile to bound the root}
}
\description{
Calculate cubic roots using the Rconics package
}
