% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{defcombine}
\alias{defcombine}
\title{helper function, borrowed from foreach}
\usage{
defcombine(a, ...)
}
\arguments{
\item{a}{value,}

\item{...}{existing list}
}
\description{
helper function, borrowed from foreach
}
