% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{denseMatrixToSparse}
\alias{denseMatrixToSparse}
\title{Convert matrix to a SparseM csr matrix}
\usage{
denseMatrixToSparse(m)
}
\arguments{
\item{m}{a matrix in standard, dense format}
}
\value{
A matrix in SparseM csr format
}
\description{
Convert matrix to a SparseM csr matrix
}
