% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{do_matched_call}
\alias{do_matched_call}
\title{run function only with arguments
that match the arguments for f}
\usage{
do_matched_call(f, ...)
}
\arguments{
\item{f}{function}

\item{...}{arguments to match}
}
\description{
run function only with arguments
that match the arguments for f
}
