% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{ensureSpecFullRank}
\alias{ensureSpecFullRank}
\title{Ensure that a regression specification is full rank}
\usage{
ensureSpecFullRank(spec_mat, col_names)
}
\arguments{
\item{spec_mat}{An M by N regression specification matrix, where M > N.}

\item{col_names}{Column names of the regression specification.}
}
\value{
A full-rank specification matrix. If the input is full-rank, returns the
input unmodified. Otherwise, returns a matrix with a subset of the columns
from the input.
}
\description{
Ensure that a regression specification is full rank
}
\details{
Verifies if a regression specification is full-rank. If the input
is rank-deficient, identifies and drops columns so that the remaining
matrix is full-rank.
}
