% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_helpers.R
\name{eval_density_R}
\alias{eval_density_R}
\title{A function to evaluate the quantile or density of
given data based on normal distribution}
\usage{
eval_density_R(y, alphas, m = 1, s = 1, distn = "p", tails = "gaussian")
}
\arguments{
\item{y}{a vector of quantile or pdf}

\item{alphas}{a vector of specifed quantiles to interpolate}

\item{m}{mean of the normal distribution}

\item{s}{standard deviation of the normal distribution}

\item{distn}{a string which indicates what type of distribution to return. See details.}

\item{tails}{what distribution to use for the tails}
}
\value{
A vector of quantiles or density corresponding to y
}
\description{
A function to evaluate the quantile or density of
given data based on normal distribution
}
\details{
the `distn` argument takes on "p" to evaluate a PF, "c" to evaluate a CDF,
and "q" to evaluate a quantile distribution.
}
