% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_helpers.R
\name{eval_PDF}
\alias{eval_PDF}
\title{Evaluate PDF given quantiles and residuals}
\usage{
eval_PDF(y, quantiles, alphas, distn = "p", tails = "gaussian")
}
\arguments{
\item{y}{a vector of quantile or pdf}

\item{quantiles}{a matrix of size N b p made up of the previously fitted quantiles}

\item{alphas}{which quantiles were previously computed, vector of probs}

\item{distn}{a string which indicates what type of distribution to return. See details.}

\item{tails}{what distribution to use when computing the tails}
}
\description{
Evaluate PDF given quantiles and residuals
}
\details{
the `distn` argument takes on "p" to evaluate a PF, "c" to evaluate a CDF,
and "q" to evaluate a quantile distribution.
}
