% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{fe_calc_admm}
\alias{fe_calc_admm}
\title{Given Gamma, recover latent factor structure via principal components and admm}
\usage{
fe_calc_admm(
  Y,
  T_vec,
  J_vec,
  N,
  J,
  Time,
  Gamma = NULL,
  tau = 0.5,
  rho = 0.1,
  init_k = 1,
  desired_dim = NULL,
  err_u_thresh = 0.01,
  err_g_thresh = 0.01,
  verbose = F,
  maxiter = 100,
  min_N = 30
)
}
\arguments{
\item{Y}{vector of outcomes}

\item{T_vec}{vector of time assignment}

\item{J_vec}{vector of group assignments}

\item{N}{number of assets}

\item{J}{integer representing number of groups}

\item{Time}{integer representing number of periods}

\item{tau}{quantile to calculate}

\item{rho}{stochastic relaxation parameter}

\item{err_u_thresh}{error threshold for u}

\item{min_N}{minimum N for group-wise OLS}
}
\description{
Given Gamma, recover latent factor structure via principal components and admm
}
