% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{findLastRedundantCol}
\alias{findLastRedundantCol}
\title{Grab the last colinear column in a matrix}
\usage{
findLastRedundantCol(x)
}
\arguments{
\item{x}{matrix to drop columns from}
}
\description{
Grab the last colinear column in a matrix
}
