% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lasso.R
\name{fit_lasso}
\alias{fit_lasso}
\title{Fit a quantile regression w/ a lasso penalty}
\usage{
fit_lasso(
  x,
  y,
  tau = 0.5,
  lambda = NULL,
  weights = NULL,
  intercept = TRUE,
  coef.cutoff = 1e-04,
  method = "sfn",
  ...
)
}
\arguments{
\item{x}{design matrix}

\item{y}{outcome variable}

\item{tau}{target quantile}

\item{lambda}{weight for penalization factor}

\item{weights}{optional observation weights}

\item{intercept}{whether or not to model the intercept}

\item{coef.cutoff}{what value for a coefficient is considered 0}

\item{method}{what underlying regression method to use for fitting}

\item{...}{other arguments to pass to method}
}
\description{
Fit a quantile regression w/ a lasso penalty
}
