% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_penalize_approx_quantile_model}
\alias{fit_penalize_approx_quantile_model}
\title{Compute quantile regression via accelerated gradient descent using
Huber approximation, warm start based on data subset}
\usage{
fit_penalize_approx_quantile_model(
  X,
  y,
  X_sub,
  y_sub,
  tau,
  init_beta,
  mu = 1e-15,
  maxiter = 100000L,
  beta_tol = 1e-04,
  check_tol = 1e-06,
  intercept = 1L,
  num_samples = 1000,
  warm_start = 1L,
  scale = 1L
)
}
\arguments{
\item{X}{design matrix}

\item{y}{outcome vector}

\item{X_sub}{subset of X matrix to use for "warm start" regression}

\item{y_sub}{subset of y to use for "warm start" regression}

\item{tau}{target quantile}

\item{init_beta}{initial guess at beta}

\item{mu}{neighborhood over which to smooth}

\item{maxiter}{maximum number of iterations to run}

\item{beta_tol}{tolerance for largest element of gradient, used
for early stopping}

\item{check_tol}{loss function change tolerance for early stopping}

\item{intercept}{location of the intercept column, using R's indexing}

\item{num_samples}{number of samples used for subset of matrix used for warm start}

\item{warm_start}{integer indicating whether to "warm up" on a subsample
of the data}

\item{scale}{whether to scale x & y variables}
}
\description{
Compute quantile regression via accelerated gradient descent using
Huber approximation, warm start based on data subset
}
