% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{fitQuantileRegression}
\alias{fitQuantileRegression}
\title{Estimate a single quantile regression}
\usage{
fitQuantileRegression(X, y, tau, algorithm = "rq.fit.sfn_start_val", ...)
}
\arguments{
\item{X}{specification matrix for X variables}

\item{y}{outcome variable}

\item{tau}{quantile to regress}

\item{algorithm}{which algorithm to use}

\item{...}{other arguments to be passed to the algorithm}
}
\description{
Estimate a single quantile regression
}
