% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{gamma_ols_iter}
\alias{gamma_ols_iter}
\title{Iterate ols given Principal components regression data}
\usage{
gamma_ols_iter(e, est, w, reg_data, J_vec, min_N, ...)
}
\arguments{
\item{e}{errors, ignored}

\item{est}{previous estimate, ignored}

\item{w}{weight vector}

\item{reg_data}{regression data from principal components}

\item{J_vec}{group vector to split ols}

\item{min_N}{minimum threshold for observations}
}
\description{
Iterate ols given Principal components regression data
}
