% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_helpers.R
\name{get_intercept}
\alias{get_intercept}
\title{Finds which column of X has the intercept term}
\usage{
get_intercept(X)
}
\arguments{
\item{X}{design matrix for regression}
}
\description{
Finds which column of X has the intercept term
}
