% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_effects.R
\name{get_marginal_effects}
\alias{get_marginal_effects}
\title{Calculates the marginal effects  of an N x p matrix (wide-format) of qreg coefficients}
\usage{
get_marginal_effects(
  qreg_coeffs,
  avg_spacings,
  j_star,
  calc_se = TRUE,
  qreg_vcv_vec = NULL
)
}
\arguments{
\item{qreg_coeffs}{wide-format of calculated spacings point estimates}

\item{avg_spacings}{average spacings matrix, which can be calculated from setting calculateAvgME = TRUE or using
the avg_spacing function directly}

\item{j_star}{the first quantile the user wishes to predict (usually the middle one)}

\item{calc_se}{boolean value, indicating whether the user wishes to calculate the marginal effect standard errors}

\item{qreg_vcv_vec}{variance-covariance matrix from point estimates, only necessary if calculating standard errors}
}
\value{
list of values: avgME: calculated average marginal effects
avgME_se (user-specified): standard errors on the calculated marginal effects
}
\description{
Calculates the marginal effects  of an N x p matrix (wide-format) of qreg coefficients
}
