% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.R
\name{get_ols_r_squared}
\alias{get_ols_r_squared}
\title{Get R^2 of OLS fit}
\usage{
get_ols_r_squared(a, y, betas, weight_vec = NULL)
}
\arguments{
\item{a}{Regression specification matrix}

\item{y}{Outcome vector}

\item{betas}{Parameters from OLS fit}

\item{weight_vec}{Optional vector of weights'}
}
\description{
Get R^2 of OLS fit
}
