% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{get_strata}
\alias{get_strata}
\title{Get clusters for subsampling given a formula}
\usage{
get_strata(cluster_matrix)

get_strata(cluster_matrix)
}
\arguments{
\item{cluster_matrix}{matrix of cluster column values}
}
\description{
Get clusters for subsampling given a formula

Get clusters for subsampling given a formula
}
