% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{get_underlying}
\alias{get_underlying}
\title{Get the data inside the s4 slot of this sparse matrix class}
\usage{
get_underlying(data)
}
\arguments{
\item{data}{some thing that could be a sparse matrix}
}
\description{
Get the data inside the s4 slot of this sparse matrix class
}
