% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_se.R
\name{getClusterIndices}
\alias{getClusterIndices}
\title{Draw a subsample of clusters and optionally assigns random exponential weights}
\usage{
getClusterIndices(cluster_col)
}
\arguments{
\item{cluster_col}{Column vector with list of cluster indices}
}
\value{
N_cluster x 2 vector with starting
                  and ending positions of each cluster.
                  NOTE that this requires that data are
                  sorted in this dimension
}
\description{
Draw a subsample of clusters and optionally assigns random exponential weights
}
