% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_se.R
\name{getDeltas}
\alias{getDeltas}
\title{Get coefficient Deltas to calculate Delta Method standard errors}
\usage{
getDeltas(
  spacing_coef,
  dep_col,
  data,
  alphas,
  jstar,
  weights = NULL,
  desired_coeffs = NULL,
  ...
)
}
\arguments{
\item{spacing_coef}{coefficients from spacing regression}

\item{dep_col}{Column of response variable.}

\item{data}{Regression specification matrix.}

\item{alphas}{Quantiles to be estimated.}

\item{jstar}{First quantile to be estimated (usually the center one)}

\item{weights}{vector of observations weights. Defaults to 1/N
for each observation if unspecified}

\item{...}{unused for now}

\item{desired_coefs}{which coefficients to calculate standard errors.
If want standard errors for first and second parameters for each quantile,
should desired_coeffs = c(1,2)}
}
\value{
List w/ p elements, with jth element a
       Nxk matrix corresponding to the delta_ij values
       for the jth quantile estimated
}
\description{
Get coefficient Deltas to calculate Delta Method standard errors
}
