% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_se.R
\name{getDeltaVarCov}
\alias{getDeltaVarCov}
\title{Computes variance-covariance matrix and standard errors of parameter estimates}
\usage{
getDeltaVarCov(delta, cluster_indices = NULL, weights = NULL, ...)
}
\arguments{
\item{delta}{list of length p, where each element is
an Nxk matrix of delta_ij values with jth element corresponding to jth quantile
estimated (would be obtained by running getSEDeltas function)}

\item{cluster_indices}{vector of length N corresponding to labels
of which cluster given observation belongs to (defaults to NULL)
weights--observation weights. Defaults to NULL, which corresponds
to using 1/N as the observation weights}

\item{...}{unused for now}
}
\value{
A list with two elements: varcov--kp X kp variance covariance matrix of parameters,
se--kp x 1 standard errors of parameters
}
\description{
Computes variance-covariance matrix and standard errors of parameter estimates
}
