% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{getMarginalEffects}
\alias{getMarginalEffects}
\title{Calculates the marginal effects  of an N x p matrix (wide-format) of qreg coefficients}
\usage{
getMarginalEffects(
  qreg_coeffs,
  avg_spacings,
  j_star,
  calcSE = TRUE,
  qreg_vcv_vec = NULL
)
}
\arguments{
\item{qreg_coeffs}{wide-format of calculated spacings point estimates}

\item{avg_spacings}{average spacings matrix, which can be calculated from setting calculateAvgME = TRUE or using
the calcAvgSpacing function directly}

\item{j_star}{the first quantile the user wishes to predict (usually the middle one)}

\item{calcSE}{boolean value, indicating whether the user wishes to calculate the marginal effect standard errors}

\item{qreg_vcv_vec}{variance-covariance matrix from point estimates, only necessary if calculating standard errors}
}
\value{
list of values: avgME: calculated average marginal effects
avgME_se (user-specified): standard errors on the calculated marginal effects
}
\description{
Calculates the marginal effects  of an N x p matrix (wide-format) of qreg coefficients
}
