% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{getRank}
\alias{getRank}
\title{Computes the rank of a matrix. Outputs consistent results across
various matrix types.}
\usage{
getRank(m, TOL = 1e-09)
}
\arguments{
\item{m}{A matrix, m}

\item{TOL}{tolerance for rank calculation}
}
\value{
The rank of the matrix m
}
\description{
Computes the rank of a matrix. Outputs consistent results across
various matrix types.
}
