% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{getRows}
\alias{getRows}
\title{Function which gets resampled rows}
\usage{
getRows(rows, sampling_method, ...)
}
\arguments{
\item{rows}{rows to resample}

\item{sampling_method}{method for resampling}

\item{...}{other arguments passed on to the method}
}
\description{
Function which gets resampled rows
}
