% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{getWeights}
\alias{getWeights}
\title{Function that returns the correct weights for weighted bootstrap}
\usage{
getWeights(weight_vec, draw_weights, n)
}
\arguments{
\item{weight_vec}{vector of existing regression weights}

\item{draw_weights}{whether draw random exponential weights}

\item{n}{number of weights to draw}
}
\description{
Function that returns the correct weights for weighted bootstrap
}
