% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glob_obs_mat}
\alias{glob_obs_mat}
\title{Glob observations w/ residuals above a certain magnitude}
\usage{
glob_obs_mat(X, r, thresh)
}
\arguments{
\item{X}{design matrix}

\item{r}{vector of residuals}

\item{thresh}{threshold to use when computing globs}
}
\description{
Glob observations w/ residuals above a certain magnitude
}
