% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_se.R
\name{\%dopar\%}
\alias{\%dopar\%}
\title{Local do parallel function}
\usage{
obj \%dopar\% ex
}
\arguments{
\item{obj}{\code{foreach} object used to control the evaluation
of \code{ex}.}

\item{ex}{the \code{R} expression to evaluate.}
}
\description{
Local do parallel function
}
