% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{huber_grad_descent}
\alias{huber_grad_descent}
\title{Compute quantile regression via accelerated gradient descent using
Huber approximation}
\usage{
huber_grad_descent(
  X,
  y,
  tau,
  init_beta,
  mu = 1e-04,
  maxiter = 10000L,
  beta_tol = 1e-04,
  check_tol = 1e-06,
  intercept = 1L
)
}
\arguments{
\item{X}{design matrix}

\item{y}{outcome vector}

\item{tau}{target quantile}

\item{init_beta}{initial guess at beta}

\item{mu}{neighborhood over which to smooth}

\item{maxiter}{maximum number of iterations to run}

\item{beta_tol}{tolerance for largest element of gradient, used
for early stopping}

\item{check_tol}{loss function change tolerance for early stopping}

\item{intercept}{what column contains the intercept}
}
\description{
Compute quantile regression via accelerated gradient descent using
Huber approximation
}
