% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_one_row}
\alias{interpolate_one_row}
\title{Interpolate a single set of quantiles}
\usage{
interpolate_one_row(quantile, alphas, x, distn = "q", tails = "gaussian")
}
\arguments{
\item{quantile}{matrix of fitted quantiles to interpolate}

\item{alphas}{values those quantiles are fit to}

\item{x}{values along which to evaluate the CDF, PDF, or quantile function}
}
\description{
Interpolate a single set of quantiles
}
\details{
Used as a helper function for interpolate_quantiles
}
