% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_quantiles}
\alias{interpolate_quantiles}
\alias{interpolate_quantiles.qs}
\alias{interpolate_quantiles.matrix}
\title{Interpolate quantiles and return a cumulative distribution function}
\usage{
interpolate_quantiles(object, ...)

\method{interpolate_quantiles}{qs}(
  object,
  newdata = NULL,
  grid = seq_quant(0, 1, by = 0.01),
  parallel = T,
  ncores = getCores(),
  row_thresh = 20,
  tails = "gaussian",
  distn = "q",
  ...
)

\method{interpolate_quantiles}{matrix}(
  object,
  alphas,
  grid = seq_quant(0, 1, by = 0.01),
  parallel = T,
  ncores = getCores(),
  row_thresh = 100,
  tails = "gaussian",
  distn = "q",
  ...
)
}
\arguments{
\item{object}{either matrix of quantiles or a fitted model with a relevant
method to dispatch}

\item{...}{other arguments passed to interpolate quantiles}

\item{newdata}{New data to use when calculating quantiles to interpolate}

\item{grid}{grid along which to evaluate the CDF}

\item{parallel}{whether to work in parallel}

\item{ncores}{number of cores to use}

\item{row_thresh}{required minimum number of observations to use parallel processing}

\item{tails}{what distribution to use when fitting the tails, either "gaussian" or "exponential"}

\item{distn}{what to return, q for quantile, c for cdf, p for pdf}

\item{alphas}{values those quantiles are fit to}
}
\description{
Interpolate quantiles and return a cumulative distribution function
}
