% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lasso.R
\name{lasso_cv_search}
\alias{lasso_cv_search}
\title{Search for optimal lambda via cross-validation}
\usage{
lasso_cv_search(
  x,
  y,
  tau = 0.5,
  weights = NULL,
  method = "sfn",
  intercept = TRUE,
  nfolds = 10,
  foldid = NULL,
  nlambda = 100,
  eps = 1e-04,
  init.lambda = 1,
  parallel = T,
  coef.cutoff = 1e-05,
  ...
)
}
\arguments{
\item{x}{design matrix for regression}

\item{y}{outcome variable for regression}

\item{tau}{target quantile}

\item{weights}{optional weights for regression}

\item{method}{method to be used for penalized quantile regression
(usually one of "sfn" or "br")}

\item{intercept}{Whether to model the intercept or not}

\item{nfolds}{number of folds to use for crossvalidation}

\item{foldid}{optional pre-specified fold identifier (for example,
if you want the folds to satisfy underlying data groupings)}

\item{nlambda}{number of lambdas to search over}

\item{eps}{smallest lambda used in search}

\item{init.lambda}{initial lambda for search}

\item{parallel}{whether to run cv scoring in parallel or not}

\item{...}{other parameters to pass on to fitting method}
}
\description{
Search for optimal lambda via cross-validation
}
\details{
Searches for a range of lambdas by first expanding the
max lambda if necessary, then finding the smallest lambda
that sets all coefficients to zero. Then it computes kfold CV scores
along a grid of lambdas, returning the scores and the smallest lambda.
}
