% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{leaveRows}
\alias{leaveRows}
\title{Function that doesn't reorder rows (for weighted bootstrap)}
\usage{
leaveRows(rows, ...)

leaveRows(rows, ...)
}
\arguments{
\item{rows}{rows to potentially resample}

\item{...}{other arguments, ignored}
}
\description{
Function that doesn't reorder rows (for weighted bootstrap)

Function that doesn't reorder rows (for weighted bootstrap)
}
