% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.R
\name{make_se_mat}
\alias{make_se_mat}
\title{Make matrix into a "standard error" matrix}
\usage{
make_se_mat(mat, d)
}
\arguments{
\item{mat}{matrix to pass}

\item{d}{number of significant digits}
}
\description{
Make matrix into a "standard error" matrix
}
