% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helpers.R
\name{makePlan}
\alias{makePlan}
\title{Make a plan for the future parallel backend}
\usage{
makePlan(ncores)
}
\arguments{
\item{ncores}{number of cores}
}
\description{
Make a plan for the future parallel backend
}
\details{
This function gets the number of cores set by user
and subsequently makes a plan that is operating system dependent.
If it is on windows, it uses multisession since forks are not an option.
Otherwise it uses multicore. If you would like to pass a custom plan.
}
