% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{map_parallel}
\alias{map_parallel}
\title{Map a function along a list in parallel}
\usage{
map_parallel(
  l,
  f,
  ...,
  parallel = T,
  ncores = getCores(),
  thresh = 20,
  collapse = "list"
)
}
\arguments{
\item{l}{list whose items are passed as data to f}

\item{f}{function to map along rows}

\item{...}{parameters passed to function}

\item{ncores}{number of cores to use}

\item{thresh}{required length of list to use parallel processing}

\item{collapse}{what function to use when collapsing arguments}
}
\description{
Map a function along a list in parallel
}
\details{
only works if the function's dependencies are completely
contained in "quantspace" package
}
