% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{map_rows_parallel}
\alias{map_rows_parallel}
\title{Map a function along rows of a matrix or data.frame}
\usage{
map_rows_parallel(
  mat,
  f,
  ...,
  parallel = T,
  ncores = getCores(),
  row_thresh = 20,
  collapse = "rbind"
)
}
\arguments{
\item{mat}{matrix or data.frame whose rows are passed as data to f}

\item{f}{function to map along rows}

\item{...}{parameters passed to function}

\item{parallel}{T/F, whether to operate in parallel or in sequence}

\item{ncores}{number of cores to use}

\item{row_thresh}{required number of rows to use parallel processing}

\item{collapse}{function to use when collapsing list of objects}
}
\description{
Map a function along rows of a matrix or data.frame
}
\details{
only works if the function's dependencies are completely
contained in "quantspace" package
}
