% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{matrixTocsc}
\alias{matrixTocsc}
\title{Convert Matrix Sparse Matrix to SparseM row compressed matrix}
\usage{
matrixTocsc(X)
}
\arguments{
\item{X}{column compressed matrix from Matrix package}
}
\description{
Convert Matrix Sparse Matrix to SparseM row compressed matrix
}
