% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{matrixTocsr}
\alias{matrixTocsr}
\title{Convert Matrix Sparse Matrix to SparseM row compressed matrix}
\usage{
matrixTocsr(X)
}
\arguments{
\item{X}{column compressed matrix from Matrix package}
}
\description{
Convert Matrix Sparse Matrix to SparseM row compressed matrix
}
