% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{melt_gamma}
\alias{melt_gamma}
\title{Reformat Gamma matrix of quantile estimates for fixed effects}
\usage{
melt_gamma(Gamma, J_levels, T_levels)
}
\arguments{
\item{Gamma}{matrix of coefficients for fixed effects}

\item{J_levels}{levels of grouping factor}

\item{T_levels}{levels of time factor}
}
\description{
Reformat Gamma matrix of quantile estimates for fixed effects
}
