% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{ols_by_group}
\alias{ols_by_group}
\title{Generate predictions from OLS by group quickly}
\usage{
ols_by_group(vec, reg_data, groups, min_N)
}
\arguments{
\item{vec}{vector of estimated Gamma values, length of Y variable}

\item{reg_data}{matrix or data.frame of principal components, with one row for each gamma, one column for each desired latent factor}

\item{groups}{vector of group variables to use in the independent regressions}

\item{min_N}{minimum N per regression}
}
\description{
Generate predictions from OLS by group quickly
}
