% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{plot.distributional_effects_list}
\alias{plot.distributional_effects_list}
\title{Visualize distributional effects}
\usage{
\method{plot}{distributional_effects_list}(distributional_effects_list, what = "cdf", tail_level = 0.01, ...)
}
\arguments{
\item{what}{what to plot, one of "pdf", "cdf", "quantiles"}

\item{tail_level}{probability level to stop plotting tail}

\item{...}{other arguments, ignored for now}

\item{distributional_effects}{estimate of distributional effects}
}
\description{
Visualize distributional effects
}
