% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{post_processed_grad_descent}
\alias{post_processed_grad_descent}
\title{Smoothed Quantile Regression with Post-Processing}
\usage{
post_processed_grad_descent(
  X,
  y,
  tau,
  lambda = 0,
  nwarmup_samples = 0.1 * nrow(X),
  lp_size = 10000,
  intercept = NULL
)
}
\arguments{
\item{X}{design matrix}

\item{y}{outcome variable}

\item{tau}{target quantile}

\item{lambda}{optional weight on penalty function}

\item{nwarmup_samples}{number of samples to use for warmup in approximat
quantile regression}

\item{lp_size}{size of linear programming problem passed to the simplex
algorithm}
}
\description{
Smoothed Quantile Regression with Post-Processing
}
\details{
This function performs smoothed quantile regression w/ post-processing
to ensure accuracy of the approximate first-order method.
}
