% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{predict.group_ols}
\alias{predict.group_ols}
\title{"Predict" from group ols}
\usage{
\method{predict}{group_ols}(est, DT, ...)
}
\arguments{
\item{DT}{data.table, unused here}

\item{...}{other parameters, ignored}

\item{original}{estimate}
}
\description{
"Predict" from group ols
}
