% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.R
\name{predict.qs}
\alias{predict.qs}
\title{Predict quantiles given fitted spacings model}
\usage{
\method{predict}{qs}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{fitted \code{qs} model}

\item{newdata}{optional new data to pass}

\item{...}{additional arguments, ignored for now}
}
\value{
This function returns a set of predictive quantiles
for each data point, at the quantiles where spacings were
estimated.
}
\description{
Predict quantiles given fitted spacings model
}
