% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{predict.two_way_fe_estimate}
\alias{predict.two_way_fe_estimate}
\title{Prediction function for two-way fixed effects given DT}
\usage{
\method{predict}{two_way_fe_estimate}(est, DT, ...)
}
\arguments{
\item{est}{previous estmiate of two way quantile fixed effects}

\item{DT}{data.table that contains design matrix columns to allow prediction}
}
\description{
Prediction function for two-way fixed effects given DT
}
\details{
Right now this is super fragile, expecting specifically
one column named "Group" and one named "Year". This is not meant to be viewed
externally to the package.
}
