% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_effects.R
\name{print.qs_me}
\alias{print.qs_me}
\title{Print method for quantspace marginal effects}
\usage{
\method{print}{qs_me}(x, ...)
}
\arguments{
\item{x}{marginal effects to print}

\item{...}{other arguments for s3 consistency, ignored for now}
}
\description{
Print method for quantspace marginal effects
}
