% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printWarnings}
\alias{printWarnings}
\title{Generate warning messages for regression model}
\usage{
printWarnings(model)
}
\arguments{
\item{model}{result of call to regressResiduals}
}
\description{
Generate warning messages for regression model
}
