% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_helpers.R
\name{q_spline_R}
\alias{q_spline_R}
\title{Computes the tail parameters and the second derivatives
given quantiles}
\usage{
q_spline_R(quantiles, alphas, tails = "gaussian")
}
\arguments{
\item{quantiles}{a matrix of size N by p containing the
fitted quantiles}

\item{alphas}{quantiles that were computed}

\item{tails}{what distribution to use for the tails}
}
\value{
A list containing the tail parameters and the second
derivatives
}
\description{
Computes the tail parameters and the second derivatives
given quantiles
}
