% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qr_drop_colinear_columns}
\alias{qr_drop_colinear_columns}
\title{Drop Colinear Columns from dense matrix}
\usage{
qr_drop_colinear_columns(X, tol = 1e-09)
}
\arguments{
\item{X}{matrix to drop colinear columns from}

\item{tol}{tolerance for when pivot is zero in rank calculations}
}
\description{
Drop Colinear Columns from dense matrix
}
