% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.R
\name{qs}
\alias{qs}
\title{Compute quantile regressions via quantile spacings}
\usage{
qs(
  formula,
  data,
  quantiles = c(0.9, 0.75, 0.5, 0.25, 0.1),
  baseline_quantile = 0.5,
  se_method = "delta",
  ...
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can
be coerced to that class): a symbolic description of the model
to be fitted.}

\item{data}{an optional data frame, list or environment (or object
coercible by as.data.frame to a data frame) containing the variables in the model.
If not found in data, the variables are taken from environment(formula)}

\item{quantiles}{vector of quantiles to be estimated}

\item{baseline_quantile}{baseline quantile to measure spacings from (defaults to 0.5)}

\item{se_method}{one of "delta" or "boot". Delta uses the delta method approximation,
while boot bootstraps the standard errors.}
}
\description{
Compute quantile regressions via quantile spacings
}
