% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.R
\name{qs_control}
\alias{qs_control}
\title{Control quantreg_spacing parameters}
\usage{
qs_control(
  trunc = TRUE,
  small = NULL,
  lambda = NULL,
  output_quantiles = TRUE,
  calc_avg_me = FALSE,
  ...
)
}
\arguments{
\item{trunc}{whether to truncate residual values below the argument "small"}

\item{small}{level of "small" values to guarentee numerical stability. If not specified, set dynamically based on the standard deviation of the outcome variable.}

\item{lambda}{For penalized regression, you can specify a level of lambda which will weight the penalty. If not set, will be determined based on 10-fold cross-validation.}

\item{output_quantiles}{whether to save fitted quantiles as part of the function output}

\item{calc_avg_me}{whether to return average marginal effects as part of the fitted object}

\item{...}{Other arguments, ignored for now}
}
\description{
Control quantreg_spacing parameters
}
\details{
control parameters to pass to the control arguments of \code{\link{quantreg_spacing}},
the lower-level function called by \code{\link{qs}}. This is set via the function \code{\link{qs_control}},
which returns a named list, with elements including:
}
