% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_helpers.R
\name{quad_form}
\alias{quad_form}
\title{Quadratic form of the cubic polynomial}
\usage{
quad_form(a, b, c, discriminant = NULL)
}
\arguments{
\item{a}{parameter 1 in a cubic equation}

\item{b}{parameter 2 in a cubic equation}

\item{c}{parameter 3 in a cubic equation}

\item{discriminant}{discriminant of the polynomial}
}
\description{
Quadratic form of the cubic polynomial
}
