% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{quantreg_spacing}
\alias{quantreg_spacing}
\title{Lower level function which calculates the quantile spacing regression coefficients}
\usage{
quantreg_spacing(
  y,
  X,
  var_names,
  alpha,
  jstar,
  algorithm = "rq.fit.sfn",
  weights = NULL,
  control = list(small = 1e-06, trunc = TRUE, start_list = NA, output_quantiles =
    FALSE, calc_avg_me = FALSE),
  ...
)
}
\arguments{
\item{y}{Column of response variable.}

\item{X}{Regression specification matrix.}

\item{var_names}{RHS regression variable names.}

\item{alpha}{Quantiles to be estimated.}

\item{jstar}{First quantile to be estimated (usually the center one)}

\item{algorithm}{The name of a function which will estimate a quantile regression.
Defaults to rq.fit.sfn_start_val. Must be a string, as it is passed to \code{do.call}}

\item{weights}{vector of optional weights}

\item{control}{control parameters to pass to the control arguments of \code{\link{quantreg_spacing}},
the lower-level function called by \code{\link{qs}}. This is set via the function \code{\link{qs_control}},
which returns a named list, with elements including:
\itemize{
\item \code{trunc}: whether to truncate residual values below the argument "small"
\item \code{small}: level of "small" values to guarentee numerical stability. If not specified, set dynamically based on the standard deviation of the outcome variable.
\item \code{output_quantiles}: whether to save fitted quantiles as part of the function output
\item \code{calc_avg_me}: whether to return average marginal effects as part of the fitted object
\item \code{lambda}: the penalization factor to be passed to penalized regression algorithms
}}

\item{...}{other parameters passed to the algorithm}

\item{small}{Minimum size of residuals for computational accuracy.}

\item{trunc}{Boolean value; if true, replace those dependent values less than small with small itself;
else, only use rows with residuals greater than small}

\item{start_list}{Starting values for regression optimization.}
}
\value{
Returns a list of coefficients.
num_betas is an x by p matrix of estimated parameters for each supplied quantiles.
pseudo_r is a  1 by p matrix of psuedo R^2 values for each quantile estimate.
warnings is a 1 by p matrix of warnings produced by each quantile regression call.
iter: is a 1 by p matrix of iterations ran by each quantile regression call.
}
\description{
Lower level function which calculates the quantile spacing regression coefficients
}
