% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{quantRegress}
\alias{quantRegress}
\title{Runs quantile regression on residuals of the model (calculates spaces around jstar quantile)}
\usage{
quantRegress(
  reg_spec_data,
  ehat,
  sv,
  ind_hat,
  tau,
  trunc,
  small,
  control,
  weight_vec = NULL
)
}
\arguments{
\item{reg_spec_data}{result of ensureSpecRank function; regression matrix with full rank}

\item{ehat}{current residuals; subset of which to be used as dependent column}

\item{sv}{starting values (can be NA's) to be passed to sfn_start_val function}

\item{ind_hat}{column vector indicating which rows to be used in quantile regression}

\item{tau}{estimated quantile}

\item{trunc}{Boolean value; if true, replace those dependent values less than small with small itself;
else, only use rows with residuals greater than small}

\item{small}{Value used with trunc; values less than small 'blow up' too greatly when logged}

\item{control}{control list to be passed to sfn_start_val function}

\item{weight_vec}{vector of optional weights}
}
\value{
List of estimated coefficients, warnings, iterations, and controls as in
standard quantile regression function
}
\description{
Runs quantile regression on residuals of the model (calculates spaces around jstar quantile)
}
