% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{quantRegSpacing}
\alias{quantRegSpacing}
\title{Computes coefficients for the quantile regression spacing method.}
\usage{
quantRegSpacing(
  dep_col,
  data,
  var_names,
  alpha,
  jstar,
  small = 0.001,
  trunc = FALSE,
  start_list,
  weight_vec = NULL
)
}
\arguments{
\item{dep_col}{Column of response variable.}

\item{data}{Regression specification matrix.}

\item{var_names}{RHS regression variable names.}

\item{alpha}{Quantiles to be estimated.}

\item{jstar}{First quantile to be estimated (usually the center one)}

\item{small}{Minimum size of residuals for computational accuracy.}

\item{trunc}{Boolean value; if true, replace those dependent values less than small with small itself;
else, only use rows with residuals greater than small}

\item{start_list}{Starting values for regression optimization.}

\item{weight_vec}{vector of optional weights}

\item{p}{Length of alpha.}
}
\value{
list of coef: num_betas x p matrix of estimated parameters for each supplied quantiles,
        pseudo_r: 1 x p matrix of psuedo R^2 values for each quantile estimate,
        warnings: 1 x p matrix of warnings produced by each quantile regression call,
        iter: 1 x p matrix of iterations ran by each quantile regression call
}
\description{
Computes coefficients for the quantile regression spacing method.
}
