% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lasso.R
\name{randomly_assign}
\alias{randomly_assign}
\title{Randomly assign fold ids}
\usage{
randomly_assign(n, nfolds)
}
\arguments{
\item{n}{number of observations to assign}

\item{nfolds}{number of folds to assign}
}
\description{
Randomly assign fold ids
}
\details{
This is used internally for assignment of observations
for kfold cross-validation
}
