% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{regressResiduals}
\alias{regressResiduals}
\title{Runs quantile regression on residuals of the model (calculates spaces around jstar quantile)}
\usage{
regressResiduals(
  reg_spec_data,
  ehat,
  ind_hat,
  tau,
  trunc,
  small,
  weight_vec,
  algorithm = "rq.fit.sfn_start_val",
  ...
)
}
\arguments{
\item{reg_spec_data}{result of ensureSpecRank function; regression matrix with full rank}

\item{ehat}{current residuals; subset of which to be used as dependent column}

\item{ind_hat}{column vector indicating which rows to be used in quantile regression}

\item{tau}{estimated quantile}

\item{trunc}{Boolean value; if true, replace those dependent values less than small with small itself;
else, only use rows with residuals greater than small}

\item{small}{Value used with trunc; values less than small 'blow up' too greatly when logged}

\item{weight_vec}{vector of optional weights}

\item{algorithm}{The name of a function which will estimate a quantile regression.
Defaults to rq.fit.sfn_start_val. Must be a string, as it is passed to `do.call`}

\item{...}{other arguments to the function specified by the algorithm argument}
}
\value{
List of estimated coefficients, warnings, iterations, and controls as in
standard quantile regression function
}
\description{
Runs quantile regression on residuals of the model (calculates spaces around jstar quantile)
}
