% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{reorder_coefficients}
\alias{reorder_coefficients}
\title{Reorder coefficients in case intercept wasn't the first term}
\usage{
reorder_coefficients(coefficients, intercept)
}
\arguments{
\item{coefficients}{vector of coefficients}

\item{intercept}{column number for intercept term}
}
\description{
Reorder coefficients in case intercept wasn't the first term
}
