% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{replace_nas}
\alias{replace_nas}
\title{Replace all NA values in a data.table}
\usage{
replace_nas(DT)
}
\arguments{
\item{DT}{data.table to put 0 values in (instead of NA values)}
}
\description{
Replace all NA values in a data.table
}
\details{
edits values by reference, so no need to return
or assign things. I know, not very functional programming,
but it is really fast.
credit: https://stackoverflow.com/questions/7235657/fastest-way-to-replace-nas-in-a-large-data-table
}
