% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{repMat}
\alias{repMat}
\title{For copying matrices as in Matlab (works for sparse matrices)}
\usage{
repMat(X, m, n)
}
\arguments{
\item{X}{matrix to replicate}

\item{m}{number of times replicate the matrix rows}

\item{n}{number of times replicate the matrix columns}
}
\description{
For copying matrices as in Matlab (works for sparse matrices)
}
