% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{rescale_coefficients}
\alias{rescale_coefficients}
\title{Rescale coefficients estimated on scaled data to match unscaled data}
\usage{
rescale_coefficients(coefs, mu_x, sigma_x, intercept)
}
\arguments{
\item{coefs}{scaled coefficients}

\item{mu_x}{vector of means for all (non-intercept) X variables}

\item{sigma_x}{vector of scales for all (non-intercept) X variables}

\item{intercept}{an integer for which column in design matrix is the intercept}
}
\description{
Rescale coefficients estimated on scaled data to match unscaled data
}
