% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.R
\name{residuals.qs}
\alias{residuals.qs}
\title{Method for getting residuals from fitted qs model}
\usage{
\method{residuals}{qs}(object, ...)
}
\arguments{
\item{object}{fitted \code{qs} model}

\item{...}{additional arguments, ignored for now}
}
\description{
Method for getting residuals from fitted qs model
}
