% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{rho}
\alias{rho}
\title{Function for calculating Pseudo-R^2 values
Evaluates the check objective function (possibly weighted) for QREG}
\usage{
rho(u, tau = 0.5, weight_vec = NULL)
}
\arguments{
\item{u}{vector of residuals}

\item{tau}{probability index parameter (quantile of interest)}

\item{weight_vec}{optional vector of weights}
}
\value{
Sum of absolute residuals
}
\description{
Function for calculating Pseudo-R^2 values
Evaluates the check objective function (possibly weighted) for QREG
}
