% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.R
\name{round_if}
\alias{round_if}
\title{Round if x is numeric, otherwise don't}
\usage{
round_if(df, d)
}
\arguments{
\item{df}{data.frame whose columns I want to round}

\item{d}{number of digits}
}
\description{
Round if x is numeric, otherwise don't
}
