% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{rq.fit.br}
\alias{rq.fit.br}
\title{Version that complies with more general requirements}
\usage{
rq.fit.br(X, y, tau = 0.5, weight_vec = NULL, control, ...)
}
\arguments{
\item{X}{structure of the design matrix X stored in csr format}

\item{y}{response vector}

\item{tau}{target quantile}

\item{weight_vec}{optional vector of weights}

\item{control}{ignored}

\item{...}{additional quantities passed to rq.fit.br}
}
\description{
Version that complies with more general requirements
}
