% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{rq.fit.lasso}
\alias{rq.fit.lasso}
\title{Quantile Regression w/ Lasso Penalty}
\usage{
rq.fit.lasso(
  X,
  y,
  tau,
  lambda,
  weights,
  scale_x = T,
  method = "agd",
  nfold = 10,
  nlambda = 50,
  parallel = F,
  ...
)
}
\arguments{
\item{X}{Design matrix, X}

\item{y}{outcome variable, y}

\item{tau}{quantile to estimate}

\item{lambda}{penalty parameter}

\item{weights}{optional vector of weights}

\item{scale_x}{whether to scale the design matrix before estimation}

\item{method}{method to use when fitting underlying quantile regression algorithm}

\item{nfold}{number of folds to use when cross-validating}

\item{nlambda}{number of lambdas to search over.}

\item{parallel}{whether to run cv search in parallel, if applicable}

\item{...}{other arguments to pass to underlying fitting algorithm}
}
\description{
Quantile Regression w/ Lasso Penalty
}
