% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{rq.fit.sfn_start_val}
\alias{rq.fit.sfn_start_val}
\title{Sparse Regression Quantile Fitting with Weights}
\usage{
rq.fit.sfn_start_val(
  X,
  y,
  tau = 0.5,
  rhs = (1 - tau) * c(t(a) \%*\% rep(1, length(y))),
  control,
  sv,
  weights = NULL,
  lambda,
  ...
)
}
\arguments{
\item{X}{structure of the design matrix X stored in csr format}

\item{y}{outcome vector}

\item{tau}{desired quantile}

\item{rhs}{the right-hand-side of the dual problem; regular users shouldn't need to specify this,
but in special cases can be quite usefully altered to meet special needs.
See e.g. Section 6.8 of Koenker (2005).}

\item{control}{control parameters for fitting routines: see \code{\link[quantreg:sfn.control]{quantreg::sfn.control()}}}

\item{sv}{startin value for optimization, useful when bootstrapping}

\item{weights}{Optional vector of weights for regression}

\item{lambda}{ignored}

\item{...}{other parameters, ignored}
}
\description{
Sparse Regression Quantile Fitting with Weights
}
\details{
A wrapper around the rq.fit.sfn function from the quantreg package,
extended to allow for  a user-supplied starting value and weights
}
