% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{rq.fit.two_pass}
\alias{rq.fit.two_pass}
\title{Quantile regression approximated w/ huber loss followed by post-processing}
\usage{
rq.fit.two_pass(
  X,
  y,
  tau = 0.5,
  weights = NULL,
  control,
  lambda,
  n_samples = min(c(ceiling(nrow(X)/10), 10000)),
  ...
)
}
\arguments{
\item{X}{design matrix}

\item{y}{outcome vector}

\item{tau}{target quantile}

\item{weights}{optional weight vector}

\item{control}{ignored for now}

\item{lambda}{ignored for now}

\item{n_samples}{number of observations to use in "warmup" regression}

\item{...}{other arguments, ignored for now}

\item{smoothing_window}{neighborhood around 0 which is
smoothed by either typical least squares or appropriately
tilted least squares loss function}

\item{beta_tol}{stopping rule based on max value of gradient}

\item{check_tol}{stopping rule based on change in the loss function}

\item{maxiter}{largest number of iterations allowed}

\item{init_beta}{initial guess at betas}

\item{scale}{whether to scale x and y variables in regression}

\item{intercept}{optional integer indicating intercept column
that identifies initial values}
}
\description{
Quantile regression approximated w/ huber loss followed by post-processing
}
