% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_helpers.R
\name{scale_for_lasso}
\alias{scale_for_lasso}
\title{Scale matrix for lasso regression}
\usage{
scale_for_lasso(X, intercept)
}
\arguments{
\item{X}{design matrix}

\item{intercept}{column number for intercept}
}
\description{
Scale matrix for lasso regression
}
