% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{seq_quant}
\alias{seq_quant}
\title{For sequencing along probabilities}
\usage{
seq_quant(from, to, by)
}
\arguments{
\item{from}{where to start}

\item{to}{where to end}

\item{by}{how far should each step be}
}
\description{
For sequencing along probabilities
}
