% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_helpers.R
\name{spDiag}
\alias{spDiag}
\title{Create sparse diagonal matrix with vector x on diagonal}
\usage{
spDiag(v)
}
\arguments{
\item{v}{Vector to use as the diagonal of the matrix}
}
\description{
Create sparse diagonal matrix with vector x on diagonal
}
