% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_helpers.R
\name{splint_R}
\alias{splint_R}
\title{A function to conduct the interpolation given data and fitted quantiles}
\usage{
splint_R(
  y,
  quantiles,
  alphas,
  y2,
  tail_param_u,
  tail_param_l,
  q_shift,
  q_stretch,
  tails = "gaussian",
  distn = "c"
)
}
\arguments{
\item{y}{a vector of quantile or pdf}

\item{quantiles}{a matrix of size N b p}

\item{alphas}{quantiles that were computed}

\item{y2}{second derivatives}

\item{tail_param_u}{the tail parameters for upper quantiles}

\item{tail_param_l}{the tail parameters for lower quantiles}

\item{q_shift}{normalization parameters}

\item{q_stretch}{normalization parameters}

\item{tails}{based on what distribution to compute the tails}

\item{distn}{a string which indicates what type of distribution to return. See details.}
}
\value{
A vector of quantiles or density corresponding to y
}
\description{
A function to conduct the interpolation given data and fitted quantiles
}
\details{
the \code{distn} argument takes on "p" to evaluate a PF, "c" to evaluate a CDF,
and "q" to evaluate a quantile distribution.
}
