% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_se.R
\name{subsampleStandardErrors}
\alias{subsampleStandardErrors}
\title{Computes standard errors for the quantile regression spacing method using
subsampling.}
\usage{
subsampleStandardErrors(
  dep_col,
  data,
  algorithm,
  var_names,
  alpha,
  jstar,
  cluster_indices = NULL,
  stratum_indices = NULL,
  M = 1,
  draw_weights = TRUE,
  num_bs = 100,
  parallel = F,
  num_cores = getCores(),
  small = 1e-06,
  trunc = FALSE,
  start_model,
  weight_vec = NULL,
  ...
)
}
\arguments{
\item{dep_col}{Column of response variable.}

\item{data}{Regression specification matrix.}

\item{algorithm}{function which is actually used to fit each quantile regression}

\item{var_names}{RHS regression variable names.}

\item{alpha}{Quantiles to be estimated.}

\item{jstar}{First quantile to be estimated (usually the center one)}

\item{cluster_indices}{N_cluster x 2 vector with starting
and ending positions of each cluster.
NOTE that this requires that data are
sorted in this dimension}

\item{stratum_indices}{N_cluster x 1 vector with stratum indices
(factor variables). This enables the
researcher to stratify (e.g., on cluster size)
Set to NULL to take a random subsample}

\item{M}{percentage of clusters to sample, must be greater than 0 but less than or equal to 1}

\item{draw_weights}{Boolean value; if true, draw a vector of exponential
weights to use in subsample}

\item{num_bs}{Number of subsample draws (must be greater than 1).}

\item{parallel}{whether to run in parallel or not}

\item{num_cores}{number of cores to use, defaults to option set by `options(mc.cores)` if not specified}

\item{small}{Minimum size of residuals for computational accuracy.}

\item{trunc}{Boolean value; if true, replace those dependent values less than small with small itself;
else, only use rows with residuals greater than small}

\item{start_model}{Starting values for regression's optimization.}

\item{weight_vec}{vector of same length and order as dependent column, to be used as weights for estimation
(note, if draw weights is set to TRUE, this variable will be the element-wise product
of itself and a random vector of weights)}

\item{...}{other arguments passed to quantile fitting function}
}
\value{
list of cov: num_betas x num_betas covariance matrix using
                bootstrapped subsampling covariances
          ols_cov: VCV matrix from OLS estimates
          warnings: num_bs x p matrix of warning messages
                     produced in each bootstrap's quantile regression call
         iter: num_bs x p matrix of iterations ran by each
                bootstrapped quantile regression call
         OLS: OLS point estimates
         counts: number of observations in each subsample
         coef_boot: full set of boostrapped quantile spacing coefficients
}
\description{
Computes standard errors for the quantile regression spacing method using
subsampling.
}
