% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.R
\name{summary.qs}
\alias{summary.qs}
\title{creates a table of summary output for a qs object}
\usage{
\method{summary}{qs}(object, ...)
}
\arguments{
\item{object}{an object of class qs}

\item{...}{other arguments to pass to summary}
}
\description{
creates a table of summary output for a qs object
}
