% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summary_output}
\alias{summary_output}
\title{Generates summary output for quantile and OLS fits}
\usage{
summary_output(quant_fit, ols_fit, se, alpha, varnames, ols_r_squared)
}
\arguments{
\item{quant_fit}{Named rows of quantile beta estimates.}

\item{ols_fit}{Named rows of OLS beta estimates.}

\item{se}{Vector containing covariance matrices for both quantile and OLS fits.}

\item{alpha}{Column vector representing quantiles being estimated.}

\item{varnames}{vector of variable names for output}

\item{ols_r_squared}{R-squared for the OLS fit.}
}
\value{
Summary output matrix
}
\description{
Generates summary output for quantile and OLS fits
}
