% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{update_gamma}
\alias{update_gamma}
\title{Update gamma given an initial matrix}
\usage{
update_gamma(Gamma, tune, W, E, nu, err_g_thresh = 0.01)
}
\arguments{
\item{Gamma}{Parameter matrix}

\item{tune}{Tuning parameter to pass along}

\item{W}{Number of groups at a given time}

\item{E}{Proportion of observed values at a given time (correction for imbalanced panel)}

\item{nu}{Penalization parameter used in optimization}
}
\description{
Update gamma given an initial matrix
}
