% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{vec_splint_R}
\alias{vec_splint_R}
\title{Simple vectorized version of splint_R}
\usage{
vec_splint_R(y, ...)
}
\arguments{
\item{y}{passed to splint_R}

\item{...}{other things passed to splint_R}
}
\description{
Simple vectorized version of splint_R
}
\details{
This is an internal helper function
}
