% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{warm_huber_grad_descent}
\alias{warm_huber_grad_descent}
\title{Compute quantile regression via accelerated gradient descent using
Huber approximation, warm start based on data subset}
\usage{
warm_huber_grad_descent(
  X,
  y,
  X_sub,
  y_sub,
  tau,
  init_beta,
  mu = 1e-04,
  maxiter = 10000L,
  beta_tol = 1e-04,
  check_tol = 1e-06,
  intercept = 1L,
  num_samples = 1000
)
}
\arguments{
\item{X}{design matrix}

\item{y}{outcome vector}

\item{X_sub}{subset of X matrix to use for "warm start" regression}

\item{y_sub}{subset of y to use for "warm start" regression}

\item{tau}{target quantile}

\item{init_beta}{initial guess at beta}

\item{mu}{neighborhood over which to smooth}

\item{maxiter}{maximum number of iterations to run}

\item{beta_tol}{tolerance for largest element of gradient, used
for early stopping}

\item{check_tol}{loss function change tolerance for early stopping}

\item{intercept}{location of the intercept column, using R's indexing}

\item{num_samples}{number of samples used for subset of matrix used for warm start}
}
\description{
Compute quantile regression via accelerated gradient descent using
Huber approximation, warm start based on data subset
}
