% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{distributional_effects}
\alias{distributional_effects}
\alias{distributional_effects.qs}
\alias{distributional_effects.matrix}
\title{Calculate distributional effects}
\usage{
distributional_effects(object, tails = "gaussian", ...)

\method{distributional_effects}{qs}(object, tails = "gaussian", newdata = NULL, ...)

\method{distributional_effects}{matrix}(object, tails, alphas, ...)
}
\arguments{
\item{object}{fit of class qs or matrix of fitted quantiles}

\item{tails}{one of "gaussian" or "exponential"}

\item{...}{other parameters to pass}

\item{newdata}{new data to predict distributional outcomes for}

\item{alphas}{which quantiles these were fitted at}
}
\description{
Calculate distributional effects
}
\details{
The arguments alphas and quantiles are automatically handled
if you pass an object of class "qs". If you don't pass new data to be
predicted on, it will assume that you want to calculate distributional
effects at the average of the data. This varies because of the non-linear
model for the quantile process
}
