% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{getWeights}
\alias{getWeights}
\title{Function that returns the correct weights for weighted bootstrap}
\usage{
getWeights(weights, draw_weights, cluster_index)
}
\arguments{
\item{weights}{vector of existing regression weights}

\item{draw_weights}{whether draw random exponential weights}

\item{cluster_index}{number of weights to draw}
}
\description{
Function that returns the correct weights for weighted bootstrap
}
