% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_effects.R
\name{me}
\alias{me}
\title{Get marginal effects at a set of levels for the covariates}
\usage{
me(fit, data)
}
\arguments{
\item{fit}{A fitted model from the \code{qs} function}

\item{data}{optional data.frame that specifies level of data to calculate
marginal effects}
}
\description{
Get marginal effects at a set of levels for the covariates
}
\details{
A simple function which returns
marginal effects for a given level of the dataset.
}
