% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.R
\name{ols_sparse_fit}
\alias{ols_sparse_fit}
\title{Computes OLS coefficients without losing "sparse"-ness of inputs}
\usage{
ols_sparse_fit(a, y, weight_vec = NULL)
}
\arguments{
\item{a}{Regression specification matrix}

\item{y}{Dependent variable}

\item{weight_vec}{Optional vector of weights}
}
\value{
OLS coefficients
}
\description{
Computes OLS coefficients without losing "sparse"-ness of inputs
}
