% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.R
\name{print.qs}
\alias{print.qs}
\title{Print qs summary}
\usage{
\method{print}{qs}(x, digits = 4, ...)
}
\arguments{
\item{x}{fit from qs}

\item{digits}{number of digits to print}

\item{...}{additional arguments, ignored for now}
}
\description{
Print qs summary
}
