% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{rq.fit.agd}
\alias{rq.fit.agd}
\title{Quantile Regression approximated w/ huber loss}
\usage{
rq.fit.agd(
  X,
  y,
  tau = 0.5,
  weights = NULL,
  control,
  lambda,
  smoothing_window = 1e-15,
  beta_tol = 1e-10,
  check_tol = 1e-07 * nrow(X),
  maxiter = 10000,
  n_samples = min(c(ceiling(nrow(X)/10), 10000)),
  ...
)
}
\arguments{
\item{X}{design matrix}

\item{y}{outcome vector}

\item{tau}{target quantile}

\item{weights}{optional weight vector}

\item{control}{ignored for now}

\item{lambda}{ignored for now}

\item{smoothing_window}{neighborhood around 0 which is
smoothed by either typical least squares or appropriately
tilted least squares loss function}

\item{beta_tol}{stopping rule based on max value of gradient}

\item{check_tol}{stopping rule based on change in the loss function}

\item{maxiter}{largest number of iterations allowed}

\item{n_samples}{number of observations to use in "warmup" regression
that identifies initial values}
}
\description{
Quantile Regression approximated w/ huber loss
}
