% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{rq.fit.lasso}
\alias{rq.fit.lasso}
\title{Quantile Regression w/ Lasso Penalty}
\usage{
rq.fit.lasso(
  X,
  y,
  tau,
  lambda,
  weights,
  scale_x = T,
  method = "br",
  nfold = 10,
  ...
)
}
\arguments{
\item{X}{Design matrix, X}

\item{y}{outcome variable, y}

\item{tau}{quantile to estimate}

\item{lambda}{penalty parameter}

\item{weights}{optional vector of weights}

\item{scale_x}{whether to scale the design matrix before estimation}

\item{method}{method argument to be passed to \link[quantreg:rq]{quantreg::rq}}

\item{nfold}{number of folds to use when cross-validating}

\item{...}{other arguments to pass to \link[rqPen:rq.lasso.fit]{rqPen::rq.lasso.fit}}
}
\description{
Quantile Regression w/ Lasso Penalty
}
