% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{rq.fit.lasso}
\alias{rq.fit.lasso}
\title{Quantile Regression w/ Lasso Penalty}
\usage{
rq.fit.lasso(X, y, tau, lambda, weight_vec, scale_x = T, method = "br", ...)
}
\arguments{
\item{X}{Design matrix, X}

\item{y}{outcome variable, y}

\item{tau}{quantile to estimate}

\item{lambda}{penalty parameter}

\item{weight_vec}{optional vector of weights}

\item{scale_x}{whether to scale the design matrix before estimation}

\item{...}{other arguments to pass to rqPen::rq.lasso.fit}
}
\description{
Quantile Regression w/ Lasso Penalty
}
