% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_regress.R
\name{rq.fit.two_pass}
\alias{rq.fit.two_pass}
\title{Quantile regression approximated w/ huber loss followed by post-processing}
\usage{
rq.fit.two_pass(
  X,
  y,
  tau = 0.5,
  weights = NULL,
  control,
  lambda,
  intercept = NULL,
  init_beta = NULL,
  n_samples = min(c(ceiling(nrow(X)/10), 10000)),
  ...
)
}
\arguments{
\item{X}{design matrix}

\item{y}{outcome vector}

\item{tau}{target quantile}

\item{weights}{optional weight vector}

\item{control}{ignored for now}

\item{lambda}{ignored for now}

\item{intercept}{optional integer indicating intercept column
that identifies initial values}

\item{init_beta}{initial guess at betas}

\item{n_samples}{number of observations to use in "warmup" regression}

\item{...}{other arguments, ignored for now}
}
\description{
Quantile regression approximated w/ huber loss followed by post-processing
}
