% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{resample_qs}
\alias{resample_qs}
\alias{weighted_bootstrap}
\alias{bootstrap}
\alias{subsample}
\alias{standard_errors}
\title{Computes standard errors for the quantile regression spacing method using
subsampling.}
\usage{
resample_qs(
  X,
  y,
  weights,
  sampling_method,
  alpha,
  jstar,
  control,
  algorithm,
  draw_weights,
  var_names,
  subsample_percent,
  cluster_index,
  ...
)

weighted_bootstrap(
  X,
  y,
  weights,
  sampling_method,
  alpha,
  jstar,
  control,
  algorithm,
  draw_weights,
  var_names,
  subsample_percent,
  cluster_index,
  ...
)

bootstrap(
  X,
  y,
  weights,
  sampling_method,
  alpha,
  jstar,
  control,
  algorithm,
  draw_weights,
  var_names,
  subsample_percent,
  cluster_index,
  ...
)

subsample(
  X,
  y,
  weights,
  sampling_method,
  alpha,
  jstar,
  control,
  algorithm,
  draw_weights,
  var_names,
  subsample_percent,
  cluster_index,
  ...
)

standard_errors(
  y,
  X,
  cluster_matrix,
  algorithm,
  control = qs_control(),
  std_err_control = se_control(),
  var_names,
  alpha,
  jstar,
  parallel = F,
  weights = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{X}{Regression specification matrix.}

\item{y}{Column of response variable.}

\item{weights}{vector of same length and order as dependent column, to be used as weights for estimation
(note, if draw weights is set to TRUE, this variable will be the element-wise product
of itself and a random vector of weights)}

\item{sampling_method}{One of "leaveRows", "subsampleRows", or "bootstrapRows".}

\item{alpha}{Quantiles to be estimated.}

\item{jstar}{First quantile to be estimated (usually the center one)}

\item{control}{control parameters to pass to the control arguments of \code{\link{quantreg_spacing}},
the lower-level function called by \code{\link{qs}}. This is a named list, with possible elements including:
\itemize{
\item \code{trunc}: whether to truncate residual values below the argument "small"
\item \code{small}: level of "small" values to guarentee numerical stability. If not specified, set dynamically based on the standard deviation of the outcome variable.
\item \code{lambda}: For penalized regression, you can specify a level of lambda which will weight the penalty. If not set, will be determined based on 10-fold cross-validation.
\item \code{output_quantiles}: whether to save fitted quantiles as part of the function output
\item \code{calc_avg_me}: whether to return average marginal effects as part of the fitted object
\item Any other arguments passed to specific downstream quantile regression algorithms (e.g. rq.fit).
}}

\item{algorithm}{function which is actually used to fit each quantile regression}

\item{draw_weights}{Whether to use random exponential weights for bootstrap, either TRUE or FALSE}

\item{var_names}{RHS regression variable names.}

\item{subsample_percent}{A number between 0 and one, specifying the percent of the data to subsample for standard error calculations}

\item{cluster_index}{index for clusters to sample}

\item{...}{other arguments passed to quantile fitting function}

\item{cluster_matrix}{Matrix of cluster variables, as returned by a model formula}

\item{std_err_control}{control parameters to pass to the control arguments of \code{\link{quantreg_spacing}},
the lower-level function called by \code{\link{standard_errors}}. Possible arguments include:
\itemize{
\item \code{se_method}: Method to use for standard errors, either "weighted_bootstrap",
"subsample", "bootstrap" or "custom" along with a specified subsampling method and
subsample percent. If specifying "custom", must also specify \code{subsampling_percent} and
\code{draw_weights}. If you specify "subsample", subsampling percent defaults to 0.2, but can be
changed. See details for details.
\item \code{num_bs}: Number of bootstrap iterations to use, defaults to 100.
\item \code{subsample_percent}: A number between 0 and one, specifying the percent of the data to subsample for standard error calculations
\item \code{draw_weights}: Whether to use random exponential weights for bootstrap, either TRUE or FALSE
\item \code{sampling_method} One of "leaveRows", "subsampleRows", or "bootstrapRows".
leaveRows doesn't resample rows at all. subsampleRows samples without replacement
given some percentage of the data (specified via subsample_percent), and boostrapRows
samples with replacement.`
}}

\item{parallel}{whether to run in parallel or not}

\item{seed}{Seed to be used when generating RNG}
}
\description{
Computes standard errors for the quantile regression spacing method using
subsampling.
}
