% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{resample_qs}
\alias{resample_qs}
\alias{weighted_bootstrap}
\alias{bootstrap}
\alias{subsample}
\alias{standard_errors}
\title{Computes standard errors for the quantile regression spacing method using
subsampling.}
\usage{
resample_qs(
  data,
  dep_col,
  weight_vec,
  sampling_method,
  draw_weights,
  alpha,
  jstar,
  small,
  trunc,
  algorithm,
  subsample_percent,
  var_names,
  ...
)

weighted_bootstrap(
  data,
  dep_col,
  weight_vec,
  sampling_method,
  alpha,
  jstar,
  small,
  trunc,
  algorithm,
  draw_weights,
  var_names,
  subsample_percent,
  ...
)

bootstrap(
  data,
  dep_col,
  weight_vec,
  sampling_method,
  alpha,
  jstar,
  small,
  trunc,
  algorithm,
  draw_weights,
  var_names,
  subsample_percent,
  ...
)

subsample(
  data,
  dep_col,
  sampling_method,
  alpha,
  jstar,
  small,
  trunc,
  algorithm,
  subsample_percent,
  var_names,
  draw_weights,
  weight_vec,
  ...
)

standard_errors(
  dep_col,
  data,
  cluster_matrix,
  algorithm,
  se_method,
  var_names,
  alpha,
  jstar,
  num_bs = 100,
  parallel = F,
  small = 1e-06,
  trunc = FALSE,
  weight_vec = NULL,
  subsample_percent,
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{Regression specification matrix.}

\item{dep_col}{Column of response variable.}

\item{weight_vec}{vector of same length and order as dependent column, to be used as weights for estimation
(note, if draw weights is set to TRUE, this variable will be the element-wise product
of itself and a random vector of weights)}

\item{sampling_method}{One of "leaveRows", "subsampleRows", or "bootstrapRows".
leaveRows doesn't resample rows at all. subsampleRows samples without replacement
given some percentage of the data (specified via subsample_percent), and boostrapRows
samples with replacement.}

\item{draw_weights}{Boolean value; if true, draw a vector of exponential
weights to use in subsample}

\item{alpha}{Quantiles to be estimated.}

\item{jstar}{First quantile to be estimated (usually the center one)}

\item{small}{Minimum size of residuals for computational accuracy.}

\item{trunc}{Boolean value; if true, replace those dependent values less than small with small itself;
else, only use rows with residuals greater than small}

\item{algorithm}{function which is actually used to fit each quantile regression}

\item{subsample_percent}{percentage of clusters to sample, must be greater than 0 but less than or equal to 1}

\item{var_names}{RHS regression variable names.}

\item{...}{other arguments passed to quantile fitting function}

\item{cluster_matrix}{Matrix of cluster variables, as returned by a model formula}

\item{se_method}{Method to use for standard errors, either "weighted_bootstrap",
"subsample", "bootstrap" or "resample_qs" along with a specified subsampling method
(one of "leaveRows", "subsampleRows", or "bootstrapRows") and a subsampling percent.}

\item{num_bs}{Number of subsample draws (must be greater than 1).}

\item{parallel}{whether to run in parallel or not}

\item{seed}{Seed to be used when generating RNG}
}
\description{
Computes standard errors for the quantile regression spacing method using
subsampling.
}
