% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_errors.R
\name{subsampleRows}
\alias{subsampleRows}
\title{Function that fully bootstraps rows}
\usage{
subsampleRows(rows, subsample_percent, ...)
}
\arguments{
\item{rows}{rows to potentially resample}

\item{subsample_percent}{percentage of data to use in subsampling}

\item{...}{other arguments, ignored}
}
\description{
Function that fully bootstraps rows
}
